% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmef.R
\name{bmef}
\alias{bmef}
\title{Fit the Bayesian Mixed-Effects Model for Multilevel Two-Way Functional Data (BMEF)}
\usage{
bmef(
  Y,
  X,
  JJ,
  tt,
  ff,
  R,
  K_T,
  K_F,
  n_burn,
  n_sample,
  params = NULL,
  alg_type = "BMEF-1",
  threshold = 0.05,
  save_all = FALSE
)
}
\arguments{
\item{Y}{A 3D array of functional responses of dimension \code{n × J × (TT × FF)}.
The third dimension represents the column-vectorized time-frequency matrix of dimension \code{TT × FF} for each subject-condition pair.}

\item{X}{A numeric matrix of subject-level covariates of dimension \code{n × p}.
The intercept should be included explicitly (e.g., for intercept-only, \code{p = 1}).}

\item{JJ}{A binary matrix of dimension \code{n × J}, where \code{JJ[i, j] = 1} if subject \code{i} has data observed for condition \code{j}, and \code{0} otherwise.}

\item{tt}{A numeric vector of time points defining the temporal grid.}

\item{ff}{A numeric vector of frequency points defining the frequency grid.}

\item{R}{The maximum CP rank.}

\item{K_T}{Number of marginal basis functions for the time dimension.}

\item{K_F}{Number of marginal basis functions for the frequency dimension.}

\item{n_burn}{Number of burn-in MCMC iterations.}

\item{n_sample}{Number of post-burn-in MCMC samples to retain.}

\item{params}{A list specifying prior hyperparameters. The list includes:
\itemize{
  \item{\code{Sigma2_delta}}: {A numeric vector of length \code{p}, specifying the diagonal elements of the covariance matrix \eqn{\Sigma_{\delta}} for the prior for \eqn{\delta_{j,r}}. Defaults to \code{rep(10, p)}.}
  \item{\code{a_gamma}}: {\eqn{a_{\gamma}}  for the Inverse-Gamma prior on \eqn{\sigma_{\gamma}^2}. Defaults to  \code{1}.}
  \item{\code{b_gamma}}: {\eqn{b_{\gamma}}  for the Inverse-Gamma prior on \eqn{\sigma_{\gamma}^2}. Defaults to  \code{1}.}
  \item{\code{a_omega}}: {\eqn{a_{\omega}}  for the Inverse-Gamma prior on \eqn{\sigma_{\omega_i}^2}. Defaults to  \code{1}.}
  \item{\code{b_omega}}: {\eqn{b_{\omega}}  for the Inverse-Gamma prior on \eqn{\sigma_{\omega_i}^2}. Defaults to  \code{1}.}
  \item{\code{a_delta}}: {\eqn{a_{\delta}}  for the Beta prior on \eqn{\delta_{j,r}}. Defaults to  \code{1}.}
  \item{\code{b_delta}}: {\eqn{b_{\delta}}  for the Beta prior on \eqn{\delta_{j,r}}. Defaults to  \code{1}.}
}}

\item{alg_type}{A character string specifying the algorithm type. Need to be either \code{"BMEF-1"} or \code{"BMEF-2"}.
Default is \code{"BMEF-1"}.}

\item{threshold}{A numeric threshold used to truncate \eqn{\tau_r} for rank selection. Defaults to \code{0.05}.}

\item{save_all}{Logical. If \code{TRUE}, all MCMC samples including burn-in are saved. Defaults to \code{FALSE}.}
}
\value{
A list containing the following components (with \code{K = K_T × K_F}, \code{Rs} denoting the selected rank, and \code{ns} denoting the number of saved MCMC iterations, where \code{ns = n_sample} if \code{save_all = FALSE}, and \code{ns = n_burn + n_sample} otherwise):

\item{alpha_s}{An 4D array of dimension \code{ns × K × J × n}, storing samples of the fixed effect coefficients.}
\item{gamma_s}{An 3D array of dimension \code{ns × K × n}, storing samples of the subject-level random effects (shared across conditions).}
\item{omega_s}{An 4D array of dimension \code{ns × K × J × n}, storing samples of the subject-by-condition-level random effects.}
\item{u_s}{A 3D array of dimension \code{ns × Rs × K_T}, storing samples of the factor matrix \code{U}.}
\item{v_s}{A 3D array of dimension \code{ns × Rs × K_F}, storing samples of the factor matrix \code{V}.}
\item{delta_s}{A 4D array of dimension \code{ns × J × Rs × p}, storing samples of the covariate coefficients.}
\item{Sigma2_s}{A list containing posterior samples of variance components:}
  \itemize{
    \item{\code{Sigma2_epsilon_s}}: {A numeric vector of length \code{ns}, storing samples of  \eqn{\sigma^2_\epsilon}.}
    \item{\code{Sigma2_gamma_s}}: {A numeric vector of length \code{ns}, storing samples of  \eqn{\sigma^2_\gamma}.}
    \item{\code{Sigma2_omega_s}}: {A matrix of dimension \code{ns × n}, storing samples of  \eqn{\sigma^2_{\omega_i}}.}
 }
\item{tt}{A numeric vector of time points defining the temporal grid (from the input).}
\item{ff}{A numeric vector of frequency points defining the frequency grid (from the input).}
\item{Y}{A 3D array of functional responses of dimension \code{n × J × (TT × FF)} (from the input).}
\item{X}{A numeric matrix of subject-level covariates of dimension \code{n × p}(from the input).}
}
\description{
This function fits the proposed Bayesian mixed-effects models for multilevel two-way functional data (BMEF).
Two frameworks are implemented: \code{"BMEF-1"}, which accounts for heterogeneous variances, and \code{"BMEF-2"}, which assumes homogeneous variance.
}
\details{
The inputs are multilevel two-way functional data evaluated for \code{n} subjects,
\code{J} conditions, \code{TT} time points, and \code{FF} frequency points. This implementation assumes uniform priors on the factor matrices \code{U} and \code{V}.
}
