% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmef.R
\name{inference.decompose.bmef}
\alias{inference.decompose.bmef}
\title{Posterior Inference for the Decomposition of Fixed Effects from a Fitted BMEF Model}
\usage{
inference.decompose.bmef(bmef_obj, cred_level)
}
\arguments{
\item{bmef_obj}{A fitted BMEF object returned by the function \code{bmef()}.}

\item{cred_level}{A numeric value specifying the credible level. Defaults to 0.95.}
}
\value{
A list containing the following components (with \code{TT} and \code{FF} denoting the number of time and frequency grid points used to fit BMEF, and \code{Rs} the selected CP rank of the fixed effects):
\item{base_patterns}{A list with elements \code{m}, \code{l}, and \code{u}, representing the posterior mean, lower bound, and upper bound of the credible intervals. Each element is a 3D array of size   \code{Rs} × \code{TT} × \code{FF} summarizing the two-way "base" patterns.}
\item{marginal_t}{A list with elements \code{m}, \code{l}, and \code{u}, representing the posterior mean, lower bound, and upper bound of the credible intervals. Each element is a matrix of size  \code{Rs} × \code{TT} , summarizing the marginal "principal" functions over time.}
\item{marginal_f}{A list with elements \code{m}, \code{l}, and \code{u}, representing the posterior mean, lower bound, and upper bound of the credible intervals. Each element is a matrix of size \code{Rs} × \code{FF}, summarizing the marginal "principal" functions over frequency.}
\item{weights_vals}{A list with elements \code{m}, \code{l}, and \code{u},  representing the posterior mean, lower bound, and upper bound of the credible intervals. Each element is a 3D array of size   \code{Rs} ×  \code{n} × \code{J}, summarizing the covariate-dependent weights for each subject ×  condition ×  rank combination.}
}
\description{
This function generates posterior summaries of two-way functional fixed effects using a fitted
Bayesian Mixed-Effects Model for Multilevel Two-Way Functional Data (BMEF). It returns inference results for the estimated "base" patterns,
covariate-dependent weights, and marginal "principal" functions over the time and frequency domains.
}
