% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmef.R
\name{inference.mixed.bmef}
\alias{inference.mixed.bmef}
\title{Posterior Inference for the two-way functional effects from a Fitted BMEF Model}
\usage{
inference.mixed.bmef(bmef_obj)
}
\arguments{
\item{bmef_obj}{A fitted BMEF object returned by the function \code{bmef()}.}
}
\value{
A list containing posterior predictive samples for the estimated two-way functional fixed and random effects using the posterior mean
(with \code{J} denoting the number of conditions):
\itemize{
  \item{\code{A_s}}{A 5D array of dimension \code{n × J × TT_pred × FF_pred}, storing posterior mean of the two-way functional fixed effects.}
  \item{\code{B_s}}{A 4D array of dimension \code{n × TT_pred × FF_pred}, storing posterior mean of subject-level two-way functional  random effects shared across conditions.}
  \item{\code{C_s}}{A 5D array of dimension \code{n × J × TT_pred × FF_pred}, storing posterior mean of subject-by-condition-level two-way functional random effects.}
}
}
\description{
This function generates posterior summaries of two-way functional fixed and random effects
using a fitted Bayesian Mixed-Effects Model for Multilevel Two-Way Functional Data (BMEF).
}
